添加 java go 后端测试，修复部分更改字段导致问题

+++ backend-go/internal/module/company/controller/company_controller.go
@@ -64,7 +64,7 @@ func (c *CompanyController) Disable(ctx *fiber.Ctx) error {
 	if err != nil {
 		return err
 	}
-	err = c.companyService.Disable(ctx.Context(), id)
+	err = c.companyService.Disable(ctx.Context(), uint(id))
 	if err != nil {
 		return err
 	}
